+++ src/Server/ServerCore/Systems/WorldManagement/WorldSegmentSubscriptionManager.cs
@@ -16,10 +16,10 @@
 
 using System;
 using System.Collections.Generic;
-using System.Numerics;
 using Castle.Core.Logging;
 using Sovereign.EngineCore.Components;
 using Sovereign.EngineCore.Components.Indexers;
+using Sovereign.EngineCore.Components.Types;
 using Sovereign.EngineCore.Configuration;
 using Sovereign.EngineCore.Entities;
 using Sovereign.EngineCore.Events;
@@ -56,6 +56,7 @@ public class WorldSegmentSubscriptionManager
     private readonly IEventSender eventSender;
     private readonly EntityHierarchyIndexer hierarchyIndexer;
     private readonly WorldManagementInternalController internalController;
+    private readonly KinematicComponentCollection kinematics;
     private readonly NonBlockPositionEventFilter nonBlockPositionEventFilter;
 
     /// <summary>
@@ -69,7 +70,6 @@ public class WorldSegmentSubscriptionManager
     private readonly Dictionary<GridPosition, HashSet<ulong>> playersByWorldSegments = new();
 
     private readonly PlayerPositionEventFilter positionEventFilter;
-    private readonly PositionComponentCollection positions;
 
     private readonly WorldSegmentResolver resolver;
 
@@ -88,7 +88,7 @@ public class WorldSegmentSubscriptionManager
         WorldSegmentResolver resolver, IWorldManagementConfiguration worldConfig,
         WorldSegmentActivationManager activationManager,
         IEventSender eventSender, WorldManagementInternalController internalController,
-        WorldSegmentSynchronizationManager syncManager, PositionComponentCollection positions,
+        WorldSegmentSynchronizationManager syncManager, KinematicComponentCollection kinematics,
         EntitySynchronizer synchronizer, EntityHierarchyIndexer hierarchyIndexer,
         EntityTable entityTable, NonBlockPositionEventFilter nonBlockPositionEventFilter)
     {
@@ -99,7 +99,7 @@ public class WorldSegmentSubscriptionManager
         this.eventSender = eventSender;
         this.internalController = internalController;
         this.syncManager = syncManager;
-        this.positions = positions;
+        this.kinematics = kinematics;
         this.synchronizer = synchronizer;
         this.hierarchyIndexer = hierarchyIndexer;
         this.entityTable = entityTable;
@@ -180,11 +180,11 @@ public class WorldSegmentSubscriptionManager
     ///     Called when a player position is updated.
     /// </summary>
     /// <param name="entityId">Player entity ID.</param>
-    /// <param name="position">New position.</param>
-    private void OnPlayerMoved(ulong entityId, Vector3 position)
+    /// <param name="kinematicsData">New kinematics.</param>
+    private void OnPlayerMoved(ulong entityId, Kinematics kinematicsData)
     {
         // Check if the player moved to a different world segment.
-        var center = resolver.GetWorldSegmentForPosition(position);
+        var center = resolver.GetWorldSegmentForPosition(kinematicsData.Position);
         if (currentWorldSegments[entityId] != center)
         {
             // Moved.
@@ -197,12 +197,12 @@ public class WorldSegmentSubscriptionManager
     ///     Called when a player entity is created/loaded.
     /// </summary>
     /// <param name="entityId">Player entity ID.</param>
-    /// <param name="position">New position.</param>
+    /// <param name="kinematicsData">New kinematics.</param>
     /// <param name="isLoad">Unused.</param>
-    private void OnPlayerAdded(ulong entityId, Vector3 position, bool isLoad)
+    private void OnPlayerAdded(ulong entityId, Kinematics kinematicsData, bool isLoad)
     {
         subscriptions[entityId] = new HashSet<GridPosition>();
-        var center = resolver.GetWorldSegmentForPosition(position);
+        var center = resolver.GetWorldSegmentForPosition(kinematicsData.Position);
         currentWorldSegments[entityId] = center;
         DoUpdateLogic(entityId, center);
     }
@@ -309,10 +309,10 @@ public class WorldSegmentSubscriptionManager
     private void OnNonBlockEntityAdded(ulong entityId)
     {
         // Skip if not a positioned entity.
-        if (!positions.HasComponentForEntity(entityId)) return;
+        if (!kinematics.HasComponentForEntity(entityId)) return;
 
         // Skip if there are no subscribers to the relevant world segment.
-        var segmentIndex = resolver.GetWorldSegmentForPosition(positions[entityId]);
+        var segmentIndex = resolver.GetWorldSegmentForPosition(kinematics[entityId].Position);
         var playerEntityIds = GetSubscribersForWorldSegment(segmentIndex);
         if (playerEntityIds.Count == 0) return;
 
@@ -343,8 +343,8 @@ public class WorldSegmentSubscriptionManager
     private void OnNonBlockPositionRemoved(ulong entityId, bool isUnload)
     {
         // Grab the last known world segment and cache it for eventual desync.
-        var pos = positions.GetComponentWithLookback(entityId);
-        var segmentIndex = resolver.GetWorldSegmentForPosition(pos);
+        var kinematicsData = kinematics.GetComponentWithLookback(entityId);
+        var segmentIndex = resolver.GetWorldSegmentForPosition(kinematicsData.Position);
         pendingDesyncs[entityId] = segmentIndex;
     }
 }